Reshuffle Code and Start of New Partial Structures
PathSearch broken into smaller sections for readability
MapProcessor redesign in progress
Can now determine if a ring opening reaction has taken place
